source/requirements.txt
requests==2.31.0
tenacity==8.2.3
urllib3==2.1.0
clickhouse-driver==0.2.6
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
redis==5.0.1
python-dotenv==1.0.0
